auth guards added

+++ src/app/components/login/login.component.ts
@@ -2,6 +2,7 @@ import { Component, ElementRef, EventEmitter, Output, ViewChild, OnInit } from '
 import { Router } from '@angular/router';
 import { AxiosError } from 'axios';
 import { UserDTO } from 'src/common/models/userDTO.interface';
+import { JwtService } from 'src/common/services/jwt.service';
 import { RequestService } from 'src/common/services/request.service';
 import Swal from 'sweetalert2';
 
@@ -11,18 +12,20 @@ import Swal from 'sweetalert2';
   styleUrls: ['./login.component.scss']
 })
 export class LoginComponent implements OnInit {
-  constructor(private requestService: RequestService, private router: Router) { }
+  constructor(private requestService: RequestService, private router: Router, private jwtService: JwtService) { }
   @ViewChild('username') username!: ElementRef;
   @ViewChild('password') password!: ElementRef;
   @Output() userUpdate = new EventEmitter<string>();
+  @Output() signOut = new EventEmitter<void>();
 
   async ngOnInit(): Promise<void> {
-    if (localStorage.length > 0) {
+    if (localStorage.length > 0 && await this.jwtService.verifyToken()) {
       try {
         this.userUpdate.emit(await this.requestService.getUserGroup(localStorage.getItem('givenName') as string));
         this.router.navigate(['/live']);
       }
       catch (err) {
+        console.log(err);
         localStorage.clear();
       }
     }
@@ -49,11 +52,6 @@ export class LoginComponent implements OnInit {
           icon: "success",
           text: "Signed in successfully"
         });
-        // await Swal.fire({
-        //   title: "login successful",
-        //   text: `welcome ${this.username.nativeElement.value}`,
-        //   icon: "success",
-        // });
         localStorage.setItem('accessToken', res.accessToken as string);
         localStorage.setItem('refreshToken', res.refreshToken as string);
         localStorage.setItem('userPrincipalName', res.userPrincipalName);
